move model fitls to own folder
